CST338 Project 1
This project showcases two fundamental concepts of OOP: 
Inheritance - In which a child class "inherits" the mutators and accessors of the parent class, and any other public methods implemented in the parent class. 
Polymorphism - The "IS-A" relationship between two objects or things, such as "A square IS A shape", that exists through inheritance. With polymorphism, a child object can be defined using the parent's data type. For example, if we had a parent Shape and child Square, we could instantiate the following Square object: "Shape square = new Square();", which enables the Square object to use any methods defined in the parent Shape as its own.